info_back/requirements.txt
Django==3.0.8
djangorestframework~=3.11.0
djangorestframework-simplejwt==3.2.2
django-model-utils==4.0.0
requests~=2.28.2
pip~=23.0.1
build~=0.10.0
wheel~=0.36.2
keyring~=23.13.1
pytz~=2023.3
sqlparse~=0.4.4
asgiref~=3.6.0
PyYAML~=5.4.1
docutils~=0.16
setuptools~=49.2.1
PyJWT~=2.7.0
django-environ==0.10.0